chore: split OrderedLayoutView into per-layout views, drop untested demos
Keep only the sections asserted by HorizontalLayoutViewIT and
VerticalLayoutViewIT (7 each); the removed FlexLayout, Scroller and
positioning demos had no assertions or are covered by OrderedLayoutIT
and ScrollerIT.
